+++ apps/router-e2e/__e2e__/native-navigation/app/modals/index.tsx
@@ -1,6 +1,7 @@
 import { Link, usePathname, type Href } from 'expo-router';
 import React from 'react';
 import { Text, Pressable, ScrollView, View } from 'react-native';
+import { featureFlags } from 'react-native-screens';
 
 const HomeIndex = () => {
   const pathname = usePathname();
@@ -11,13 +12,16 @@ const HomeIndex = () => {
       contentContainerStyle={{ alignItems: 'center', gap: 16 }}
       contentInsetAdjustmentBehavior="automatic">
       <View>
-        <Text>Native navigation - Index</Text>
+        <Text>Modals</Text>
         <Text>Current Path: {pathname}</Text>
+        <Text>
+          Synchronous updates:{' '}
+          {featureFlags.experiment.synchronousScreenUpdatesEnabled ? 'Enabled' : 'Disabled'}
+        </Text>
       </View>
-      <CaseLink href="/tabs" text="Native Tabs" />
-      <CaseLink href="/toolbar" text="Toolbar" />
-      <CaseLink href="/header-items" text="Header Items" />
-      <CaseLink href="/modals" text="Modals" />
+      <CaseLink href="/modals/form-sheet" text="Form Sheet Modal" />
+      <CaseLink href="/modals/form-sheet-content" text="Form Sheet Modal - fitToContent" />
+      <CaseLink href="/modals/page-sheet" text="Page Sheet Modal" />
     </ScrollView>
   );
 };